+++ CHANGELOG.md
@@ -23,6 +23,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `Grid2` is now a subclass of `Rect`
 - `Grid2.__str__()` now includes step and origin values instead of just the rectangle coordiantes
+- `Grid2.steps()` now yields coordinates even if the X or Y of the step value is 0, in which case
+  the corresponding value of the grid's origin coordinate is used for that value of the yielded
+  coordinates
 - `util.snap_num()` parameter `mult` is now typed as `float`
 - `util.snap_num()` return type is now `float`
 - `util.snap_num()` parameter `snap_fn` is now optional (defaults to the built-in `round` function)